requirements.txt
click==8.1.7
pytest==7.4.3
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 